+++ laktory/models/compute/__init__.py
@@ -1,3 +1,5 @@
 from laktory.models.compute.initscript import InitScript
 from laktory.models.compute.cluster import Cluster
 from laktory.models.compute.warehouse import Warehouse
+from laktory.models.compute.notebook import Notebook
+from laktory.models.compute.pipeline import Pipeline