+++ laktory/models/compute/notebook.py
@@ -1,15 +1,17 @@
 import os
 from typing import Any
+from typing import Literal
 from pydantic import model_validator
 from laktory.models.base import BaseModel
 from laktory.models.resources import Resources
 from laktory.models.permission import Permission
 
 
-class InitScript(BaseModel, Resources):
+class Notebook(BaseModel, Resources):
     source: str
-    dirpath: str = "/init_scripts/"
+    dirpath: str = None
     path: str = None
+    language: Literal["SCALA", "PYTHON", "SQL", "R"] = None
     permissions: list[Permission] = []
 
     @property
@@ -31,6 +33,9 @@ class InitScript(BaseModel, Resources):
             if self.dirpath:
                 self.path = f"{self.dirpath}{self.filename}"
 
+            elif "/notebooks/" in self.source:
+                self.path = "/" + self.source.split("/notebooks/")[-1]
+
         return self
 
     # ----------------------------------------------------------------------- #
@@ -38,5 +43,5 @@ class InitScript(BaseModel, Resources):
     # ----------------------------------------------------------------------- #
 
     def deploy_with_pulumi(self, name=None, groups=None, opts=None):
-        from laktory.resourcesengines.pulumi.initscript import PulumiInitScript
-        return PulumiInitScript(name=name, init_script=self, opts=opts)
+        from laktory.resourcesengines.pulumi.notebook import PulumiNotebook
+        return PulumiNotebook(name=name, notebook=self, opts=opts)